src/requirements.txt
gdown==5.2.0
layoutparser==0.3.4
matplotlib==3.7.1
mlflow==2.13.2
numpy==1.23.5
opencv_python_headless==4.9.0.80
pandas==1.5.3
pycocotools==2.0.7
PyYAML==6.0.1
torch==2.3.1
git+https://github.com/facebookresearch/detectron2

